Add ability learning logic
Two new handlers were introduced to the player controller to allow the
controllers to keep track of the object that the player is looking at.
These handlers were used in the default controller to allow the player to
learn abilitties from objects with a Teacher component.

For now learning an ability involves clicking while looking at it.
Eventually the teacher should also ensure that the player has met prereqs
and has no skill conflicts with the antireqs of the new skill before
teaching it.

Whenever we need to add a new skill we must create a new enum in Skills.cs
and then create a public static variable to list the prereqs/antireqs.