Implement interactable object behaviour
Created an abstract class to contain all types of handlers for
interactable objects. For now interactables only need to be able to
respond to selection. As a simple example I added an initial
implementation of pushable. This will prompt the PlayerManager to switch
to a PushController on the next frame.