Fix logic in player controller
The handlers that were being called in player manager (the ones that the
player controllers implement) were not exactly what we wanted. It would
make most sense to have a handler called every time the object the player
looks at changes (OnTargetChange) and one for everytime the player stares
at the same object for an extended period of time (OnHover). This solves
issue #9.

This also addresses the remarks of PR #3.