Implement SkillTree structure
The PlayerManager now has a reference to a SkillTree. The SkillTree
abstracts away the gory details of skill dependencies into a simple
interface. For now I am assuming that all prereqs must be met before you
can learn a skill and that learning a akill makes the antireqs
unlearnable.

I also took the time to move test scenes into a separate folder so that we
can keep the testing grounds apart from the actual game. The TestSceneBase
can be used as a fresh canvas to add objects to if you ned to test
features. It should not be overwritten!