Added ranged interaction, with fireball example
Developers can now create objects and projectiles and attach scripts which
are integrated with the Teacher and Interactable system to objects to make
them destructible (or otherwise interactable). Refer to the scene
FireProjectileTest.unity for an example.

A fireball prefab (aka an orange sphere) is also available to use.

Extend RangedInteraction in a new script and attach it to your target
object to take advantage of the Rangednteraction system. Also ensure that
your target has the script ProjectileTargetController. Any projectiles
should the script ProjectileController added.