Implement spawn points
Added a prefab for spawn points. When the player passes over spawn points
the current spawn point will be updated. When the player is reset (ie on
death) the players position is reset to the spawn point's. Also added a
death trigger to test. Important to note that the players starting spawn
point will now have to be set in the editor.