Implement wall climbing
Added a climbing controller. To use attach a the climbing trigger script
to an object with a trigger. Position the trigger to cover the area you
want to be climbable. So far this supports click to mount/dismount, hold
to move along the wall as well as climb off the wall to dismount/climb up.