Fix change level script
The script was not correctly setting the spawn point of the player in the
new scene. This was due to Unity not loading the scene as soon as Load is
called (it only queues the load). As a workaround we can make a static
request to the playermanager which can be fulfilled in the start method of
the player.